# docs/specs/api-contract.yaml
...
openapi: 3.0.3
info:
  title: ORBIT Backend API
  version: 1.0.0
  description: >
    Contract for the ORBIT FastAPI backend. Dashboard and Hermes are both
    clients of this API. The API exposes state and actions only — never rules;
    all conditional/triggered logic lives in n8n. Every state-changing endpoint
    writes an audit_log entry attributed to the calling actor (user | hermes | n8n).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /devices:
    get:
      tags: [devices]
      summary: List devices
      parameters:
        - { name: room_id, in: query, schema: { type: string, format: uuid } }
        - { name: type, in: query, schema: { type: string } }
        - { name: online, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Device list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Device' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
    post:
      tags: [devices]
      summary: Register a device (manual registration only — no auto-discovery)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceCreate' }
      responses:
        '201':
          description: Created. Response includes one-time MQTT credentials.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceCreated' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /devices/{deviceId}:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    get:
      tags: [devices]
      summary: Get a device (includes capabilities and current state snapshot)
      responses:
        '200':
          description: Device
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [devices]
      summary: Update device metadata (name, room, type, firmware)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceUpdate' }
      responses:
        '200':
          description: Updated device
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [devices]
      summary: Deregister a device (revokes its MQTT credentials)
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------- capabilities ----
  /devices/{deviceId}/capabilities:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    get:
      tags: [capabilities]
      summary: List a device's capabilities with current state
      responses:
        '200':
          description: Capabilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CapabilityState' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [capabilities]
      summary: Declare a new capability on a device
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapabilityCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capability' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /hermes/conversations:
    get:
      tags: [hermes]
      summary: List a user's conversations (Hermes internal)
      description: >
        Backing store for Hermes's own GET /conversations (hermes-chat-api.yaml
        proxies this). Restricted to the hermes service principal.
      parameters:
        - { name: user_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Conversations, most recent activity first
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/HermesConversation' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
        '403':
          description: Caller is not the hermes service principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    post:
      tags: [hermes]
      summary: Create a conversation (Hermes internal)
      description: >
        Called by Hermes when /chat arrives without a conversation_id
        (hermes-chat-api.yaml: "Hermes creates one via the backend"). Restricted
        to the hermes service principal.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HermesConversationCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HermesConversation' }
        '403':
          description: Caller is not the hermes service principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /hermes/conversations/{conversationId}/messages:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [hermes]
      summary: Persist one chat turn's messages (Hermes internal)
      description: >
        Hermes's message-persistence write path — hermes-tool-manifest.json
        tool `save_conversation_turn` (internal:true, never exposed to the
        LLM). Hermes persists nothing locally; every turn's messages (user
        message, tool messages, assistant reply, in order) are written here
        as a batch. Restricted to the hermes service principal
        (actor_type=hermes). Updates the conversation's last_message_at.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HermesMessagesWrite' }
      responses:
        '201':
          description: Messages persisted; ids returned in request order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HermesMessagesCreated' }
        '403':
          description: Caller is not the hermes service principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [hermes]
      summary: Message history for one conversation (Hermes internal)
      description: >
        Backing store for Hermes's GET /conversations/{id}/messages and for
        rebuilding LLM context. Restricted to the hermes service principal.
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: before, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Messages ascending by created_at
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/HermesMessage' }
        '403':
          description: Caller is not the hermes service principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Users authenticate via /auth/login. Hermes and n8n use long-lived
        service principals (JWTs with actor_type claims 'hermes' / 'n8n')
        issued from backend config — the audit log reads the actor from
        this claim.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: device_offline }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    Device:
      type: object
      properties:
        id: { type: string, format: uuid }
        room_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        type: { type: string }
        firmware_version: { type: string, nullable: true }
        is_online: { type: boolean }
        last_seen: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DeviceDetail:
      allOf:
        - $ref: '#/components/schemas/Device'
        - type: object
          properties:
            capabilities:
              type: array
              items: { $ref: '#/components/schemas/CapabilityState' }
    DeviceCreate:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string }
        room_id: { type: string, format: uuid, nullable: true }
        firmware_version: { type: string }
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/CapabilityCreate' }
    DeviceCreated:
      allOf:
        - $ref: '#/components/schemas/DeviceDetail'
        - type: object
          properties:
            mqtt_credentials:
              type: object
              description: Shown once at registration; flash onto the device.
              properties:
                username: { type: string, example: dev-4f1c... }
                password: { type: string }
                state_topic_prefix: { type: string, example: "orbit/devices/{id}/" }
    DeviceUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        room_id: { type: string, format: uuid, nullable: true }
        firmware_version: { type: string }
    Capability:
      type: object
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        capability: { type: string, example: temperature }
        data_type: { type: string, enum: [float, int, bool, string, enum, json] }
        unit: { type: string, nullable: true, example: "°C" }
        access: { type: string, enum: [read, write, read_write] }
        label: { type: string, example: Temperature }
        config: { type: object, additionalProperties: true }
    CapabilityCreate:
      type: object
      required: [capability, data_type, access, label]
      properties:
        capability: { type: string }
        data_type: { type: string, enum: [float, int, bool, string, enum, json] }
        unit: { type: string }
        access: { type: string, enum: [read, write, read_write] }
        label: { type: string }
        config: { type: object, additionalProperties: true }
    CapabilityState:
      allOf:
        - $ref: '#/components/schemas/Capability'
        - type: object
          properties:
            value:
              description: Latest reported value (type per data_type); null if never reported
              nullable: true
            reported_at: { type: string, format: date-time, nullable: true }
...
    HermesMessageCreate:
      type: object
      required: [role, channel, content]
      properties:
        role: { type: string, enum: [user, assistant, system, tool] }
        channel: { type: string, enum: [dashboard, telegram] }
        content: { type: string }
        tool_calls:
          type: array
          nullable: true
          description: >
            Backend REST tool calls made during this turn (assistant/tool
            messages only) — stored in hermes_messages.tool_calls.
          items: { type: object, additionalProperties: true }
    HermesMessagesWrite:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          minItems: 1
          description: One chat turn's messages, in conversation order
          items: { $ref: '#/components/schemas/HermesMessageCreate' }
    HermesMessagesCreated:
      type: object
      properties:
        conversation_id: { type: string, format: uuid }
        message_ids:
          type: array
          description: Created hermes_messages ids, same order as the request
          items: { type: string, format: uuid }
        last_message_at: { type: string, format: date-time }
    HermesConversation:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        title: { type: string, nullable: true }
        started_at: { type: string, format: date-time }
        last_message_at: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: true }
    HermesConversationCreate:
      type: object
      required: [user_id]
      properties:
        user_id: { type: string, format: uuid }
        title: { type: string }
        metadata: { type: object, additionalProperties: true }
    HermesMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        role: { type: string, enum: [user, assistant, system, tool] }
        channel: { type: string, enum: [dashboard, telegram] }
        content: { type: string }
        tool_calls:
          type: array
          nullable: true
          items: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }